// src/app/models/actions.model.ts
export enum ACTIONS {
    GainPower = 'Ottenere Potere',
    MoveItemOrAlly = 'Muovere un Oggetto o un Alleato',
    PlayACard = 'Giocare una Carta',
    MoveAHero = 'Muovere un Eroe',
    Activate = 'Attivare',
    Vanquish = 'Scontro',
    Fate = 'Fato',
    DiscardCards = 'Scartare',
    BrewPoison = 'Preparare Veleno'
}